Extract Strategy endpoints into WorkstationEndpoints.Strategy partial
First low-risk slice of Phase 2 (endpoint-monolith decomposition). Moves the
Strategy capability out of the 8,003-line WorkstationEndpoints.cs core into a
new sibling partial WorkstationEndpoints.Strategy.cs, mirroring the 12 existing
capability partials (Stream, SecurityMasterWorkbench, ProviderIntegrations, …).

Moved (whole methods only): MapStrategyDesignerEndpoints, MapStrategyEngineEndpoints,
and their Strategy-only helpers CreateBacktestResponse, StrategyDesignerUnavailable,
StrategyEngineUnavailable, and the nested StrategyEngineValidateRunRequest record.
Shared helpers (HasPermission, TryResolveCurrentUser, the BuildWorkstationStructuredExport
family, Operations helpers) stay in the core partial.

Because WorkstationEndpoints is a single partial class, this is a behavior-preserving
relocation: the orchestrator MapWorkstationEndpoints still calls the two Map methods
unchanged, and moved methods still reach shared private members. No signature, DI, or
route changes. Core drops 8003 -> 7680 lines; the ratchet baseline is retightened to
the new core cap.
